app_queue: Log an UNPAUSE/UNPAUSEALL if re-pausing
When an agent pauses while already paused, usually they are simply
changing their pause reason.  However, there is no UNPAUSE or UNPAUSEALL
event.  This makes reporting start and end pause spans difficult to deal
with.

For reporting sanity, this logs an UNPAUSE and UNPAUSEALL when an agent
pauses and they are already paused.